Add syntax checks, tests, and refactor ShipActionsSidebar component
